fix(review): address F04 - rename shared helper to expect-call-chain.ts to avoid collision with S5906/expect-chain.ts
